proper Maven, .gitignore & remove commons-io
