--- main/resources/.gitignore
@@ -1,1 +1,0 @@
-main/resources/c2021/
