+++ src/main/java/com/oheers/fish/competition/CompetitionType.java
