+++ src/main/java/com/oheers/fish/competition/JoinChecker.java
