+++ src/main/java/com/oheers/fish/config/CompetitionConfig.java
