+++ src/main/java/com/oheers/fish/fishing/items/Rarity.java
