+++ src/main/resources/competitions.yml
