+++ src/main/resources/fish.yml
