+++ src/main/resources/locales/messages_fr.yml
