+++ src/main/resources/rarities.yml
